# More efficient serialization of escaped characters