The following benchmark shows more than 20x speed up during serialization of string with escaped characters for these changes: 
https://github.com/plokhotnyuk/jsoniter-scala/blob/8d0ddbe182c976cd2719fc96ff00e5d2b0a4328b/jsoniter-scala-benchmark/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/macros/StringOfEscapedCharsBenchmark.scala#L82

Before:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                                  (size)   Mode  Cnt       Score       Error   Units
[info] StringOfEscapedCharsBenchmark.writeCirce                                      128  thrpt   15   16580.718 ±   489.095   ops/s
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.alloc.rate                       128  thrpt   15    1430.753 ±    42.074  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.alloc.rate.norm                  128  thrpt   15  135799.257 ±    20.702    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Eden_Space              128  thrpt   15    1432.215 ±   253.708  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Eden_Space.norm         128  thrpt   15  135806.370 ± 22262.913    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Survivor_Space          128  thrpt   15       0.021 ±     0.019  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Survivor_Space.norm     128  thrpt   15       1.955 ±     1.753    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.count                            128  thrpt   15      32.000              counts
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.time                             128  thrpt   15      34.000                  ms
```

After:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                                  (size)   Mode  Cnt       Score      Error   Units
[info] StringOfEscapedCharsBenchmark.writeCirce                                      128  thrpt   15  362014.105 ± 6127.986   ops/s
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.alloc.rate                       128  thrpt   15    1083.880 ±   18.428  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.alloc.rate.norm                  128  thrpt   15    4712.001 ±    0.001    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Eden_Space              128  thrpt   15    1069.504 ±  362.388  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Eden_Space.norm         128  thrpt   15    4651.869 ± 1581.798    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Survivor_Space          128  thrpt   15       0.015 ±    0.021  MB/sec
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.churn.PS_Survivor_Space.norm     128  thrpt   15       0.066 ±    0.092    B/op
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.count                            128  thrpt   15      24.000             counts
[info] StringOfEscapedCharsBenchmark.writeCirce:·gc.time                             128  thrpt   15      27.000                 ms
```